filename_absolute/Help_View: fix truncation to 32bit (memsize -> 32bit -> memsize)

+++ src/filename_absolute.cxx
@@ -312,7 +312,7 @@ std::string fl_filename_path_str(const std::string &filename) {
   const char *base = filename.c_str();
   const char *name = fl_filename_name(base);
   if (name) {
-    return std::string(base, (int)(name-base));
+    return std::string(base, (size_t)(name-base));
   } else {
     return std::string();
   }
